add newtow-raphson method
