+++ models/msg_queue.py
@@ -12,7 +12,7 @@ class MsgQueue(object):
     '''
 
     def __init__(self, env, server=None):
-        self.queue = deque()
+        self.queue = deque(maxlen=100)
         assert isinstance(env, simpy.Environment)
         self.__env = env
         self.__log = Logger('buffer', 'data.txt')
