+++ tests/newtow_raphson.py
@@ -3,7 +3,7 @@ l_1 = 1.0722
 l_2 = 0.48976
 s_1 = 8.4733*10**(-4)
 s_2 = 5.0201*10**(-6)
-mu = 2.181162
+mu = 2.5
 Q = np.array([[-1*s_1,s_1], [s_2,-1*s_2]])
 L = np.array([[l_1,0], [0,l_2]])
 M = np.array([[mu,0], [0,mu]])
